backend/requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
faiss-cpu
streamlit
tavily-python
langchain_ollama
langchain_community
python-dotenv
twilio
langchain-google-genai 
langchain-anthropic
langgraph-checkpoint-redis
redis
fastapi
uvicorn